Simplify web logging by providing thread to logger.

+++ src/main/java/com/drwtrading/london/reddal/workingOrders/obligations/quoting/IQuotingObligationView.java
@@ -2,7 +2,7 @@ package com.drwtrading.london.reddal.workingOrders.obligations.quoting;
 
 public interface IQuotingObligationView {
 
-    public void setRow(final String rowID, final String symbol, final int percentageOn, final boolean isStrategyOn,
+    public void setRow(final String rowID, final String symbol, final long percentageOn, final boolean isStrategyOn,
             final boolean isObligationFail);
 
     public void deleteRow(final String id);
